Artifact: Progress in artifact appendix, run experiments silently and sequentially in env test, fix formatting discrepancies

+++ artifact/test.sh
@@ -18,27 +18,37 @@ else
 	exit 1;
 fi
 
+# test that tmux works
+if ! tmux new-session -d -s artifact-tmux-test "echo 'tmux test'"; then
+	echo "The command  tmux new-session -d -s artifact-tmux-test \"echo 'tmux test'\"  failed. Please make sure that tmux is functional";
+	exit 1;
+fi
+
 # Test if data generation works. It's not detrimental if this fails, 
 # since we provide our data sets for the large experiments
 echo 'generating small data sets'
 
 rm -r experiment_data/small > /dev/null 2>&1
 if ! python3 gen_MKPID_data.py --silent experiment_data/small 2 100; then
-	echo "data set generation failed, are is the Faker package installed?"
+	echo "data set generation failed, is the Faker package installed?"
 	exit 0
 fi
 
 # do test run with one iteration on small data sets, this should not take long.
-echo 'running measurements on small data'
+echo 'Testing attacks against MK-PrivateID'
 for i in {0..49}
 do
 	l2=$((i + 64))
-	tmux new-session -d -s mpmc-chunk$i "echo '${i} ${l2}' && taskset -c ${i},${l2} python3 run_experiments.py experiment_data/small measurements/small 1 ${i}"
+	echo "cores ${i}, ${l2}"
+	# tmux new-session -d -s mpmc-chunk$i "echo '${i} ${l2}' && taskset -c ${i},${l2} python3 run_experiments.py --silent experiment_data/small measurements/small 1 ${i}"
+	taskset -c $i,$l2 python3 run_experiments.py --silent experiment_data/small measurements/small 1 $i
 done
 
+echo 'Testing attacks against PSU / PSU-CA'
 for i in {50..60}
 do
 	l2=$((i + 64))
 	cid=$((i - 50))
-	tmux new-session -d -s PSU$i "echo '${i} ${l2}' && taskset -c ${i},${l2} python3 measure_PSU.py measurements/small 100 100 1 ${cid}"
+	echo "cores ${i}, ${l2}"
+	taskset -c $i,$l2 python3 measure_PSU.py --silent measurements/small 100 100 1 $cid
 done
